pyproject.toml
[tool.poetry]
name = "user-management"
version = "0.1.0"
description = "none"
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.8"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
asyncpg = "^0.28.0"
sqlalchemy = "^2.0.23"
pydantic-settings = "^2.0.3"
pre-commit = "^3.5.0"
pyyaml = "^6.0.1"
alembic = "^1.12.1"
greenlet = "^3.0.1"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.0.1"
pydantic = {extras = ["email"], version = "^2.5.0"}
python-multipart = "^0.0.6"
redis = "^5.0.1"
httpx = "^0.25.2"
pytest = "^7.4.3"
aioboto3 = "^12.0.0"
trio = "^0.23.1"
pika = "^1.3.2"
aio-pika = "^9.4.0"

[tool.poetry.dev-dependencies]
black = "^21.10b1"
isort = "^5.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
